backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
python-dotenv
httpx
beautifulsoup4
markdown
lxml